requirements.txt
alembic==1.8.1
arabic-reshaper==2.1.4
asn1crypto==1.5.1
bcrypt==3.2.2
Brotli==1.0.9
cachelib==0.9.0
cachetools==5.2.0
canvas==0.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.1
cssselect==1.1.0
cssselect2==0.7.0
cssutils==2.5.1
cycler==0.11.0
DateTime==4.5
decouple==0.0.7
dnspython==2.2.1
dominate==2.7.0
email-to==0.1.0
email-validator==1.2.1
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.4
fpdf==1.7.2
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
pandas==1.5.1
pdfkit==1.0.0
Pillow==9.2.0
plotly==5.10.0
premailer==3.10.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.5.0
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pyparsing==3.0.9
svglib==1.4.1
tenacity==8.1.0
tinycss2==1.2.1
tk==0.1.0
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.11
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.2.1
wkhtmltopdf==0.2
WTForms==3.0.1
zope.interface==5.4.0
zopfli==0.2.1